apps/api/requirements.txt
fastapi>=0.115,<0.117
pydantic>=2.9,<3
pydantic-settings>=2.5,<3
structlog>=24.4,<25
uvicorn[standard]>=0.32,<0.33
sqlmodel>=0.0.22,<0.1
aiosqlite>=0.20,<0.21
redis>=5.1,<6
arq>=0.26,<0.27
httpx>=0.27,<0.29
python-multipart>=0.0.12,<0.1
yt-dlp>=2024.10.7
sse-starlette>=2.1,<3
